requirements.txt
Django==1.5.2
South==0.8.2
argparse==1.2.1
dj-database-url==0.2.2
dj-static==0.0.5
django-tastypie==0.10.0
django-toolbelt==0.0.1
gunicorn==18.0
psycopg2==2.5.1
python-dateutil==2.1
python-mimeparse==0.1.4
six==1.4.1
static==0.4
wsgiref==0.1.2
